3195: Move copy to share toolbar item
